+++ Changes.md
@@ -1,6 +1,8 @@
 HEAD
 ----
 
+- New: Kiba::Common::Destinations:CSV allows to write ruby hashes to a CSV file.
+
 0.0.2
 ----